FixedRatioLayout file rename
