+++ front/app/components/LandingPages/citizen/FixedRatioLayout.ts
